meshanything_train/miche/requirements.txt
absl-py==1.4.0
accelerate==0.20.3
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
braceexpand==0.1.7
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
comm==0.1.2
configargparse==1.5.3
contourpy==1.0.7
controlnet-aux==0.0.5
cycler==0.11.0
cython==0.29.33
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.0
debugpy==1.6.6
decorator==5.1.1
deepspeed==0.8.1
defusedxml==0.7.1
deprecated==1.2.14
diffusers==0.18.2
docker-pycreds==0.4.0
einops==0.6.0
executing==1.2.0
fastapi==0.101.0
fastjsonschema==2.16.2
ffmpy==0.3.1
filelock==3.9.0
flask==2.2.3
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
gradio==3.39.0
gradio-client==0.3.0
grpcio==1.51.3
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
iopath==0.1.10
ipydatawidgets==4.3.3
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.0.3
jupyter-console==6.6.1
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.3.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lightning-utilities==0.7.1
linkify-it-py==2.0.2
lmdb==1.4.1
markdown==3.4.1
markdown-it-py==2.2.0
markupsafe==2.1.2
marshmallow==3.20.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mesh2sdf==1.1.0
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.1
oauthlib==3.2.2
objaverse==0.0.7
omegaconf==2.3.0
onnxruntime==1.15.1
opencv-contrib-python==4.8.0.74
opencv-python==4.7.0.72
orjson==3.9.2
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
platformdirs==3.0.0
plotly==5.13.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pydub==0.25.1
pygltflib==1.16.0
pygments==2.14.0
pymeshlab==2022.2.post3
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pysdf==0.1.8
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pythreejs==2.4.2
pytz==2022.7.1
pywavelets==1.4.1
pyyaml==6.0
pyzmq==25.0.0
qtconsole==5.4.0
qtpy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
rtree==1.0.1
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
semantic-version==2.10.0
send2trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==63.4.3
sh==2.0.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.13.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
traittypes==0.2.1
transformers==4.30.2
trimesh==3.18.3
triton==1.1.1
typing-extensions==4.5.0
typing-inspect==0.9.0
uc-micro-py==1.0.2
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.23.2
wandb==0.13.10
wcwidth==0.2.6
webcolors==1.12
webdataset==0.2.33
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
xatlas==0.0.7
yacs==0.1.8
yarl==1.8.2
zipp==3.14.0
# torch==1.13.1+cu116 torchvision==0.14.1+cu116 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu116
https://download.pytorch.org/whl/cu116/torch-1.13.1%2Bcu116-cp39-cp39-linux_x86_64.whl#sha256=db457a822d736013b6ffe509053001bc918bdd78fe68967b605f53984a9afac5
https://download.pytorch.org/whl/cu116/torchvision-0.14.1%2Bcu116-cp39-cp39-linux_x86_64.whl#sha256=a9fc38040e133d1779f131b4497caef830e9e699faf89cd323cd58794ffb305b
https://download.pytorch.org/whl/cu116/torchaudio-0.13.1%2Bcu116-cp39-cp39-linux_x86_64.whl#sha256=5bc0e29cb78f7c452eeb4f27029c40049770d51553bf840b4ca2edd63da289ee
# torch-cluster
https://data.pyg.org/whl/torch-1.13.0%2Bcu116/torch_cluster-1.6.1%2Bpt113cu116-cp39-cp39-linux_x86_64.whl
# torch-scatter
https://data.pyg.org/whl/torch-1.13.0%2Bcu116/torch_scatter-2.1.1%2Bpt113cu116-cp39-cp39-linux_x86_64.whl
torchmetrics==0.11.1
pytorch_lightning~=1.9.3
git+https://github.com/pyvista/fast-simplification.git
git+https://github.com/skoch9/meshplot.git
git+https://github.com/NVlabs/nvdiffrast/